clear style errors

+++ src/menu/manage_expense.py
@@ -1,19 +1,20 @@
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Message, Update
 from telegram.ext import ContextTypes
 
 import menu.callbacks as cb
 from categories import categories
-from config import create_logger, log
+from config import create_logger
+from decorators import delete_old_message, log
 from expense_manager import expense_manger
 from menu.states import AUTH, DELETE_EXPENSE, MANAGE_EXPENSE, MOVE_EXPENSE
-from menu.utils import delete_old_message, make_category_inline_menu
+from menu.utils import make_category_inline_menu
 from report_manager import get_expenses_list_with_ids
 
 logger = create_logger(__name__)
 
 
+@delete_old_message(logger)
 async def manage_expenses(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    await delete_old_message(update, context)
     keyboard = [
         [InlineKeyboardButton("Перенести", callback_data=cb.manage_move_expense)],
         [InlineKeyboardButton("Удалить", callback_data=cb.manage_delete_expense)],
@@ -22,7 +23,8 @@ async def manage_expenses(update: Update, context: ContextTypes.DEFAULT_TYPE) ->
     msg = await update.message.reply_text(
         "Что сделать с расходами", reply_markup=mark_up
     )
-    context.user_data["msg"] = int(msg.id)
+    if context.user_data:
+        context.user_data["msg"] = int(msg.id)
     return MANAGE_EXPENSE
 
 
@@ -31,14 +33,16 @@ async def move_expense_request_categories(
 ) -> int:
     query = update.callback_query
     await query.answer()
-    expenses, id_map = get_expenses_list_with_ids(
-        update.effective_user.id, False, cb.report_by_date
-    )
-    context.user_data["id_map"] = id_map
-    replay_markup = make_category_inline_menu(categories)
-    await query.edit_message_text(
-        f"{expenses}\n\nВыберете новую категорию", reply_markup=replay_markup
-    )
+    if update.effective_user:
+        expenses, id_map = get_expenses_list_with_ids(
+            update.effective_user.id, False, cb.report_by_date
+        )
+        if context.user_data:
+            context.user_data["id_map"] = id_map
+        replay_markup = make_category_inline_menu(categories)
+        await query.edit_message_text(
+            f"{expenses}\n\nВыберете новую категорию", reply_markup=replay_markup
+        )
     return MOVE_EXPENSE
 
 
@@ -47,26 +51,30 @@ async def select_new_category(
 ) -> int:
     query = update.callback_query
     await query.answer()
-    context.user_data["new_category"] = int(query.data)
-    msg = await query.edit_message_text(
-        query.message.text
-        + "\n\nВведите номер или номера расходов для удаления через запятую",
-    )
-    context.user_data["msg"] = msg.id
+    if context.user_data:
+        context.user_data["new_category"] = int(query.data)
+        msg = await query.edit_message_text(
+            query.message.text
+            + "\n\nВведите номер или номера расходов для удаления через запятую",
+        )
+        if isinstance(msg, Message):
+            context.user_data["msg"] = msg.id
     return MOVE_EXPENSE
 
 
 @log(logger)
 async def move_expense(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    id_map = context.user_data.pop("id_map")
-    new_category = context.user_data.pop("new_category")
-    ids = update.message.text.split(",")
-    for id in ids:
-        expense_manger.move_expense(id_map.get(int(id)), new_category)
-    await context.bot.delete_message(
-        update.effective_user.id, context.user_data.pop("msg")
-    )
-    await update.message.reply_text("готово")
+    if context.user_data:
+        id_map = context.user_data.pop("id_map")
+        new_category = context.user_data.pop("new_category")
+        ids = update.message.text.split(",")
+        for id in ids:
+            expense_manger.move_expense(id_map.get(int(id)), new_category)
+        if update.effective_user:
+            await context.bot.delete_message(
+                update.effective_user.id, context.user_data.pop("msg")
+            )
+        await update.message.reply_text("готово")
     return AUTH
 
 
@@ -75,25 +83,28 @@ async def delete_expense_request(
 ) -> int:
     query = update.callback_query
     await query.answer()
-    expenses, id_map = get_expenses_list_with_ids(
-        update.effective_user.id, False, cb.report_by_date
-    )
-    context.user_data["id_map"] = id_map
-    msg = await query.edit_message_text(
-        f"{expenses}\n\nВведите номер или номера расходов для удаления через запятую",
-    )
-    context.user_data["msg"] = msg.id
+    if update.effective_user and context.user_data:
+        expenses, id_map = get_expenses_list_with_ids(
+            update.effective_user.id, False, cb.report_by_date
+        )
+        context.user_data["id_map"] = id_map
+        msg = await query.edit_message_text(
+            f"{expenses}\n\nВведите номер или номера расходов для удаления через запятую",
+        )
+        if isinstance(msg, Message):
+            context.user_data["msg"] = msg.id
     return DELETE_EXPENSE
 
 
 @log(logger)
 async def delete_expense(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    id_map = context.user_data.pop("id_map")
-    ids = update.message.text.split(",")
-    for id in ids:
-        expense_manger.del_expense(id_map.get(int(id)))
-    await context.bot.delete_message(
-        update.effective_user.id, context.user_data.pop("msg")
-    )
-    await update.message.reply_text("готово")
+    if context.user_data and update.effective_user:
+        id_map = context.user_data.pop("id_map")
+        ids = update.message.text.split(",")
+        for id in ids:
+            expense_manger.del_expense(id_map.get(int(id)))
+        await context.bot.delete_message(
+            update.effective_user.id, context.user_data.pop("msg")
+        )
+        await update.message.reply_text("готово")
     return AUTH
